feat(c-api) Move `module` into its own module.
