+++ lib/c-api/src/wasm_c_api/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod engine;
 pub mod externals;
+pub mod module;
 pub mod store;
 pub(crate) mod utils;
 pub mod value;
@@ -16,6 +17,7 @@ use externals::function::wasm_func_t;
 use externals::global::wasm_global_t;
 use externals::memory::wasm_memory_t;
 use externals::table::wasm_table_t;
+use module::wasm_module_t;
 use std::convert::{TryFrom, TryInto};
 use std::mem;
 use std::ptr::{self, NonNull};
@@ -26,7 +28,7 @@ use thiserror::Error;
 use value::wasm_valkind_t;
 use wasmer::{
     ExportType, Extern, ExternType, FunctionType, GlobalType, ImportType, Instance, MemoryType,
-    Module, Mutability, Pages, RuntimeError, Store, TableType, ValType,
+    Mutability, Pages, RuntimeError, TableType, ValType,
 };
 #[cfg(feature = "jit")]
 use wasmer_engine_jit::JIT;
@@ -126,113 +128,6 @@ pub unsafe extern "C" fn wasm_instance_exports(
     out.data = extern_vec.as_mut_ptr();
     // TODO: double check that the destructor will work correctly here
     mem::forget(extern_vec);
-}
-
-#[repr(C)]
-pub struct wasm_module_t {
-    pub(crate) inner: Arc<Module>,
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_module_new(
-    store_ptr: Option<NonNull<wasm_store_t>>,
-    bytes: &wasm_byte_vec_t,
-) -> Option<Box<wasm_module_t>> {
-    // TODO: review lifetime of byte slice
-    let wasm_byte_slice: &[u8] = slice::from_raw_parts_mut(bytes.data, bytes.size);
-    let store_ptr: NonNull<Store> = store_ptr?.cast::<Store>();
-    let store = store_ptr.as_ref();
-    let module = c_try!(Module::from_binary(store, wasm_byte_slice));
-
-    Some(Box::new(wasm_module_t {
-        inner: Arc::new(module),
-    }))
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_module_delete(_module: Option<Box<wasm_module_t>>) {}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_module_exports(
-    module: &wasm_module_t,
-    out: &mut wasm_exporttype_vec_t,
-) {
-    let mut exports = module
-        .inner
-        .exports()
-        .map(Into::into)
-        .map(Box::new)
-        .map(Box::into_raw)
-        .collect::<Vec<*mut wasm_exporttype_t>>();
-
-    debug_assert_eq!(exports.len(), exports.capacity());
-    out.size = exports.len();
-    out.data = exports.as_mut_ptr();
-    mem::forget(exports);
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_module_imports(
-    module: &wasm_module_t,
-    out: &mut wasm_importtype_vec_t,
-) {
-    let mut imports = module
-        .inner
-        .imports()
-        .map(Into::into)
-        .map(Box::new)
-        .map(Box::into_raw)
-        .collect::<Vec<*mut wasm_importtype_t>>();
-
-    debug_assert_eq!(imports.len(), imports.capacity());
-    out.size = imports.len();
-    out.data = imports.as_mut_ptr();
-    mem::forget(imports);
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_module_deserialize(
-    store_ptr: Option<NonNull<wasm_store_t>>,
-    bytes: *const wasm_byte_vec_t,
-) -> Option<NonNull<wasm_module_t>> {
-    // TODO: read config from store and use that to decide which compiler to use
-
-    let byte_slice = if bytes.is_null() || (&*bytes).into_slice().is_none() {
-        // TODO: error handling here
-        return None;
-    } else {
-        (&*bytes).into_slice().unwrap()
-    };
-
-    let store_ptr: NonNull<Store> = store_ptr?.cast::<Store>();
-    let store = store_ptr.as_ref();
-    let module = c_try!(Module::deserialize(store, byte_slice));
-
-    Some(NonNull::new_unchecked(Box::into_raw(Box::new(
-        wasm_module_t {
-            inner: Arc::new(module),
-        },
-    ))))
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_module_serialize(
-    module: &wasm_module_t,
-    out_ptr: &mut wasm_byte_vec_t,
-) {
-    let mut byte_vec = match module.inner.serialize() {
-        Ok(mut byte_vec) => {
-            byte_vec.shrink_to_fit();
-            byte_vec
-        }
-        Err(_) => return,
-    };
-    // ensure we won't leak memory
-    // TODO: use `Vec::into_raw_parts` when it becomes stable
-    debug_assert_eq!(byte_vec.capacity(), byte_vec.len());
-    out_ptr.size = byte_vec.len();
-    out_ptr.data = byte_vec.as_mut_ptr();
-    mem::forget(byte_vec);
 }
 
 #[no_mangle]
